Update import paths after moving remote_function folder
