+++ sagemaker-train/src/sagemaker/train/remote_function/__init__.py
@@ -13,7 +13,7 @@
 """Defines classes and helper methods used in remote function executions."""
 from __future__ import absolute_import
 
-from sagemaker.core.remote_function.client import remote, RemoteExecutor  # noqa: F401
-from sagemaker.core.remote_function.checkpoint_location import CheckpointLocation  # noqa: F401
-from sagemaker.core.remote_function.custom_file_filter import CustomFileFilter  # noqa: F401
-from sagemaker.core.remote_function.spark_config import SparkConfig  # noqa: F401
+from sagemaker.train.remote_function.client import remote, RemoteExecutor  # noqa: F401
+from sagemaker.train.remote_function.checkpoint_location import CheckpointLocation  # noqa: F401
+from sagemaker.train.remote_function.custom_file_filter import CustomFileFilter  # noqa: F401
+from sagemaker.train.remote_function.spark_config import SparkConfig  # noqa: F401
