+++ sagemaker-train/src/sagemaker/train/remote_function/runtime_environment/bootstrap_runtime_environment.py
@@ -31,7 +31,7 @@ if __package__ is None or __package__ == "":
         get_logger,
     )
 else:
-    from sagemaker.core.remote_function.runtime_environment.runtime_environment_manager import (
+    from sagemaker.train.remote_function.runtime_environment.runtime_environment_manager import (
         RuntimeEnvironmentManager,
         _DependencySettings,
         get_logger,
